HTTP server/client configuration rework.
